creation of the distribution pallet
