+++ pallets/kylin-distribution/src/models.rs
@@ -5,26 +5,26 @@ use kylin_support::types::{
 use scale_info::TypeInfo;
 use sp_runtime::{MultiSignature, RuntimeDebug};
 
-/// A single Airdrop.
+/// A single Distribution.
 #[derive(Encode, Decode, PartialEq, Eq, Copy, Clone, TypeInfo, MaxEncodedLen)]
-pub struct Airdrop<AccountId, Balance, Moment> {
-	/// Creator of the Airdrop.
+pub struct Distribution<AccountId, Balance, Moment> {
+	/// Creator of the Distribution.
 	pub creator: AccountId,
-	/// Total funds committed to the Airdrop.
+	/// Total funds committed to the Distribution.
 	pub total_funds: Balance,
 	/// Total number of recipients
 	pub total_recipients: u32,
 	/// Amount of the `total_funds` already claimed.
 	pub claimed_funds: Balance,
-	/// Starting block of the Airdrop.
+	/// Starting block of the Distribution.
 	pub start: Option<Moment>,
 	/// The minimum time, in blocks, between recipient claims.
 	pub schedule: Moment,
-	/// Set `true` if an airdrop has been explicitly disabled.
+	/// Set `true` if an distribution has been explicitly disabled.
 	pub disabled: bool,
 }
 
-/// Funds, and related information, to be claimed by an Airdrop recipient.
+/// Funds, and related information, to be claimed by an Distribution recipient.
 #[derive(Encode, Decode, PartialEq, Eq, Copy, Clone, TypeInfo, MaxEncodedLen)]
 pub struct RecipientFund<Balance, Period> {
 	/// Total funds committed for this recipient.
@@ -37,14 +37,14 @@ pub struct RecipientFund<Balance, Period> {
 	pub funded_claim: bool,
 }
 
-/// Current State of an [`Airdrop`](Airdrop).
+/// Current State of an [`Distribution`](Distribution).
 #[derive(Debug, Encode, Decode, PartialEq, Eq, Copy, Clone, TypeInfo, MaxEncodedLen)]
-pub enum AirdropState {
-	/// The Airdrop has been created but has not started.
+pub enum DistributionState {
+	/// The Distribution has been created but has not started.
 	Created,
-	/// The Airdrop has started. Recipients can claim funds.
+	/// The Distribution has started. Recipients can claim funds.
 	Enabled,
-	/// The Airdrop has ended. Recipients can **NOT** claim funds.
+	/// The Distribution has ended. Recipients can **NOT** claim funds.
 	Disabled,
 }
