+++ pallets/kylin-distribution/src/weights.rs
@@ -1,16 +1,16 @@
 use frame_support::weights::Weight;
 
 pub trait WeightInfo {
-	fn create_airdrop() -> Weight;
+	fn create_distribution() -> Weight;
 	fn add_recipient(x: u32) -> Weight;
 	fn remove_recipient() -> Weight;
-	fn enable_airdrop() -> Weight;
-	fn disable_airdrop() -> Weight;
+	fn enable_distribution() -> Weight;
+	fn disable_distribution() -> Weight;
 	fn claim(x: u32) -> Weight;
 }
 
 impl WeightInfo for () {
-	fn create_airdrop() -> Weight {
+	fn create_distribution() -> Weight {
 		0
 	}
 
@@ -22,11 +22,11 @@ impl WeightInfo for () {
 		0
 	}
 
-	fn enable_airdrop() -> Weight {
+	fn enable_distribution() -> Weight {
 		0
 	}
 
-	fn disable_airdrop() -> Weight {
+	fn disable_distribution() -> Weight {
 		0
 	}
